+++ full-service/src/json_rpc/api_v1/wallet_api.rs
@@ -4,14 +4,17 @@
 
 use crate::{
     json_rpc::api_v1::decorated_types::{
-        JsonAddress, JsonBlock, JsonBlockContents, JsonProof, JsonSubmitResponse,
-        JsonTransactionLog, JsonTxo, StringifiedJsonTxProposal,
+        JsonAccount, JsonAddress, JsonBalanceResponse, JsonBlock, JsonBlockContents, JsonProof,
+        JsonSubmitResponse, JsonTransactionLog, JsonTxo, JsonWalletStatus,
+        StringifiedJsonTxProposal,
     },
     service::WalletService,
 };
 use mc_common::logger::global_log;
 use mc_connection::{BlockchainConnection, UserTxConnection};
 use mc_fog_report_validation::FogPubkeyResolver;
+use mc_ledger_db::Ledger;
+use mc_ledger_sync::NetworkState;
 use mc_mobilecoind_json::data_types::{JsonTx, JsonTxOut, JsonTxProposal};
 use rocket_contrib::json::Json;
 use serde::{Deserialize, Serialize};
@@ -49,30 +52,6 @@ pub enum JsonCommandRequestV1 {
     },
     get_all_addresses_by_account {
         account_id: String,
-    },
-    send_transaction {
-        account_id: String,
-        recipient_public_address: String,
-        value: String,
-        input_txo_ids: Option<Vec<String>>,
-        fee: Option<String>,
-        tombstone_block: Option<String>,
-        max_spendable_value: Option<String>,
-        comment: Option<String>,
-    },
-    build_transaction {
-        account_id: String,
-        recipient_public_address: String,
-        value: String,
-        input_txo_ids: Option<Vec<String>>,
-        fee: Option<String>,
-        tombstone_block: Option<String>,
-        max_spendable_value: Option<String>,
-    },
-    submit_transaction {
-        tx_proposal: StringifiedJsonTxProposal,
-        comment: Option<String>,
-        account_id: Option<String>,
     },
     get_all_transactions_by_account {
         account_id: String,
@@ -115,15 +94,6 @@ pub enum JsonCommandResponseV1 {
     get_all_addresses_by_account {
         address_ids: Vec<String>,
         address_map: Map<String, serde_json::Value>,
-    },
-    send_transaction {
-        transaction: JsonSubmitResponse,
-    },
-    build_transaction {
-        tx_proposal: StringifiedJsonTxProposal,
-    },
-    submit_transaction {
-        transaction: JsonSubmitResponse,
     },
     get_all_transactions_by_account {
         transaction_log_ids: Vec<String>,
@@ -218,65 +188,6 @@ where
                 address_map,
             }
         }
-        JsonCommandRequestV1::send_transaction {
-            account_id,
-            recipient_public_address,
-            value,
-            input_txo_ids,
-            fee,
-            tombstone_block,
-            max_spendable_value,
-            comment,
-        } => {
-            let transaction_details = service
-                .send_transaction(
-                    &account_id,
-                    &recipient_public_address,
-                    value,
-                    input_txo_ids.as_ref(),
-                    fee,
-                    tombstone_block,
-                    max_spendable_value,
-                    comment,
-                )
-                .map_err(format_error)?;
-            JsonCommandResponseV1::send_transaction {
-                transaction: transaction_details,
-            }
-        }
-        JsonCommandRequestV1::build_transaction {
-            account_id,
-            recipient_public_address,
-            value,
-            input_txo_ids,
-            fee,
-            tombstone_block,
-            max_spendable_value,
-        } => {
-            let tx_proposal = service
-                .build_transaction(
-                    &account_id,
-                    &recipient_public_address,
-                    value,
-                    input_txo_ids.as_ref(),
-                    fee,
-                    tombstone_block,
-                    max_spendable_value,
-                )
-                .map_err(format_error)?;
-            JsonCommandResponseV1::build_transaction {
-                tx_proposal: StringifiedJsonTxProposal::from(&tx_proposal),
-            }
-        }
-        JsonCommandRequestV1::submit_transaction {
-            tx_proposal,
-            comment,
-            account_id,
-        } => JsonCommandResponseV1::submit_transaction {
-            transaction: service
-                .submit_transaction(JsonTxProposal::try_from(&tx_proposal)?, comment, account_id)
-                .map_err(format_error)?,
-        },
         JsonCommandRequestV1::get_all_transactions_by_account { account_id } => {
             let transactions = service
                 .list_transactions(&account_id)
@@ -352,7 +263,7 @@ where
 mod e2e {
     use crate::{
         db::b58_decode,
-        json_rpc::api_test_utils::{dispatch, dispatch_expect_error, setup, wait_for_sync},
+        json_rpc::api_test_utils::{dispatch, setup, wait_for_sync},
         test_utils::add_block_to_ledger_db,
     };
     use mc_common::logger::{test_with_logger, Logger};
@@ -443,248 +354,6 @@ mod e2e {
             .as_str()
             .unwrap();
         assert_eq!(unspent, "100");
-    }
-
-    #[test_with_logger]
-    fn test_build_and_submit_transaction(logger: Logger) {
-        let mut rng: StdRng = SeedableRng::from_seed([20u8; 32]);
-        let (client, mut ledger_db, _db_ctx, network_state) = setup(&mut rng, logger.clone());
-
-        // Add an account
-        let body = json!({
-            "api_version": "2",
-            "method": "create_account",
-            "params": {
-                "name": "Alice Main Account",
-                "first_block": "0",
-            }
-        });
-        let res = dispatch(&client, body, &logger);
-        let result = res.get("result").unwrap();
-        let account_obj = result.get("account").unwrap();
-        let account_id = account_obj.get("account_id").unwrap().as_str().unwrap();
-        let b58_public_address = account_obj.get("main_address").unwrap().as_str().unwrap();
-        let public_address = b58_decode(b58_public_address).unwrap();
-
-        // Add a block with a txo for this address (note that value is smaller than
-        // MINIMUM_FEE)
-        add_block_to_ledger_db(
-            &mut ledger_db,
-            &vec![public_address.clone()],
-            100,
-            &vec![KeyImage::from(rng.next_u64())],
-            &mut rng,
-        );
-
-        wait_for_sync(&client, &ledger_db, &network_state, &logger);
-
-        // Create a tx proposal to ourselves
-        let body = json!({
-            "method": "build_transaction",
-            "params": {
-                "account_id": account_id,
-                "recipient_public_address": b58_public_address,
-                "value": "42",
-            }
-        });
-        // We will fail because we cannot afford the fee, which is 100000000000 pMOB
-        // (.01 MOB)
-        dispatch_expect_error(&client, body, &logger, "{\"details\":\"Error building transaction: Wallet DB Error: Insufficient funds from Txos under max_spendable_value: Max spendable value in wallet: 100, but target value: 10000000042\",\"error\":\"TransactionBuilder(WalletDb(InsufficientFundsUnderMaxSpendable(\\\"Max spendable value in wallet: 100, but target value: 10000000042\\\")))\"}".to_string());
-
-        // Add a block with significantly more MOB
-        add_block_to_ledger_db(
-            &mut ledger_db,
-            &vec![public_address],
-            100000000000000, // 100.0 MOB
-            &vec![KeyImage::from(rng.next_u64())],
-            &mut rng,
-        );
-
-        wait_for_sync(&client, &ledger_db, &network_state, &logger);
-
-        // Create a tx proposal to ourselves
-        let body = json!({
-            "method": "build_transaction",
-            "params": {
-                "account_id": account_id,
-                "recipient_public_address": b58_public_address,
-                "value": "42000000000000", // 42.0 MOB
-            }
-        });
-        let res = dispatch(&client, body, &logger);
-        let result = res.get("result").unwrap();
-        let tx_proposal = result.get("tx_proposal").unwrap();
-        let tx = tx_proposal.get("tx").unwrap();
-        let tx_prefix = tx.get("prefix").unwrap();
-
-        // Assert the fee is correct in both places
-        let prefix_fee = tx_prefix.get("fee").unwrap().as_str().unwrap();
-        let fee = tx_proposal.get("fee").unwrap();
-        // FIXME: WS-9 - Note, minimum fee does not fit into i32 - need to make sure we
-        // are not losing precision with the JsonTxProposal treating Fee as number
-        assert_eq!(fee, "10000000000");
-        assert_eq!(fee, prefix_fee);
-
-        // Transaction builder attempts to use as many inputs as we have txos
-        let inputs = tx_proposal.get("input_list").unwrap().as_array().unwrap();
-        assert_eq!(inputs.len(), 2);
-        let prefix_inputs = tx_prefix.get("inputs").unwrap().as_array().unwrap();
-        assert_eq!(prefix_inputs.len(), inputs.len());
-
-        // One destination
-        let outlays = tx_proposal.get("outlay_list").unwrap().as_array().unwrap();
-        assert_eq!(outlays.len(), 1);
-
-        // Map outlay -> tx_out, should have one entry for one outlay
-        let outlay_index_to_tx_out_index = tx_proposal
-            .get("outlay_index_to_tx_out_index")
-            .unwrap()
-            .as_array()
-            .unwrap();
-        assert_eq!(outlay_index_to_tx_out_index.len(), 1);
-
-        // Two outputs in the prefix, one for change
-        let prefix_outputs = tx_prefix.get("outputs").unwrap().as_array().unwrap();
-        assert_eq!(prefix_outputs.len(), 2);
-
-        // One outlay confirmation number for our one outlay (no receipt for change)
-        let outlay_confirmation_numbers = tx_proposal
-            .get("outlay_confirmation_numbers")
-            .unwrap()
-            .as_array()
-            .unwrap();
-        assert_eq!(outlay_confirmation_numbers.len(), 1);
-
-        // Tombstone block = ledger height (12 to start + 2 new blocks + 50 default
-        // tombstone)
-        let prefix_tombstone = tx_prefix.get("tombstone_block").unwrap();
-        assert_eq!(prefix_tombstone, "64");
-
-        // Get current balance
-        let body = json!({
-            "api_version": "2",
-            "method": "get_balance_for_account",
-            "params": {
-                "account_id": account_id,
-            }
-        });
-        let res = dispatch(&client, body, &logger);
-        let result = res.get("result").unwrap();
-        let balance_status = result.get("balance").unwrap();
-        let unspent = balance_status
-            .get("unspent_pmob")
-            .unwrap()
-            .as_str()
-            .unwrap();
-        assert_eq!(unspent, "100000000000100");
-
-        // Submit the tx_proposal
-        let body = json!({
-            "method": "submit_transaction",
-            "params": {
-                "tx_proposal": tx_proposal,
-                "account_id": account_id,
-            }
-        });
-        let res = dispatch(&client, body, &logger);
-        let result = res.get("result").unwrap();
-        let transaction_id = result
-            .get("transaction")
-            .unwrap()
-            .get("transaction_id")
-            .unwrap()
-            .as_str()
-            .unwrap();
-        // Note - we cannot test here that the transaction ID is consistent, because
-        // there is randomness in the transaction creation.
-
-        wait_for_sync(&client, &ledger_db, &network_state, &logger);
-
-        // Get balance after submission
-        let body = json!({
-            "api_version": "2",
-            "method": "get_balance_for_account",
-            "params": {
-                "account_id": account_id,
-            }
-        });
-        let res = dispatch(&client, body, &logger);
-        let result = res.get("result").unwrap();
-        let balance_status = result.get("balance").unwrap();
-        let unspent = balance_status
-            .get("unspent_pmob")
-            .unwrap()
-            .as_str()
-            .unwrap();
-        let pending = balance_status
-            .get("pending_pmob")
-            .unwrap()
-            .as_str()
-            .unwrap();
-        let spent = balance_status.get("spent_pmob").unwrap().as_str().unwrap();
-        let secreted = balance_status
-            .get("secreted_pmob")
-            .unwrap()
-            .as_str()
-            .unwrap();
-        assert_eq!(unspent, "0");
-        assert_eq!(pending, "0");
-        assert_eq!(spent, "99990000000100");
-        assert_eq!(secreted, "0");
-
-        // FIXME: FS-93 Increment ledger manually so tx lands.
-
-        // Get the transaction_id and verify it contains what we expect
-        let body = json!({
-            "method": "get_transaction",
-            "params": {
-                "transaction_log_id": transaction_id,
-            }
-        });
-        let res = dispatch(&client, body, &logger);
-        let result = res.get("result").unwrap();
-        let transaction_log = result.get("transaction").unwrap();
-        assert_eq!(
-            transaction_log.get("direction").unwrap().as_str().unwrap(),
-            "tx_direction_sent"
-        );
-        assert_eq!(
-            transaction_log.get("value_pmob").unwrap().as_str().unwrap(),
-            "42000000000000"
-        );
-        assert_eq!(
-            transaction_log
-                .get("recipient_address_id")
-                .unwrap()
-                .as_str()
-                .unwrap(),
-            b58_public_address
-        );
-        transaction_log.get("account_id").unwrap().as_str().unwrap();
-        assert_eq!(
-            transaction_log.get("fee_pmob").unwrap().as_str().unwrap(),
-            "10000000000"
-        );
-        assert_eq!(
-            transaction_log.get("status").unwrap().as_str().unwrap(),
-            "tx_status_pending"
-        );
-        assert_eq!(
-            transaction_log
-                .get("submitted_block_height")
-                .unwrap()
-                .as_str()
-                .unwrap(),
-            "14"
-        );
-        assert_eq!(
-            transaction_log
-                .get("transaction_log_id")
-                .unwrap()
-                .as_str()
-                .unwrap(),
-            transaction_id
-        );
     }
 
     #[test_with_logger]
