+++ full-service/src/json_rpc/json_rpc_response.rs
@@ -5,11 +5,18 @@
 //! API v2
 
 use crate::json_rpc::{
+    account::Account, balance::Balance, transaction_log::TransactionLog, tx_proposal::TxProposal,
+    wallet_status::WalletStatus,
+};
+
+use crate::json_rpc::{
     account::Account, account_secrets::AccountSecrets, balance::Balance,
     wallet_status::WalletStatus,
 };
 use serde::{Deserialize, Serialize};
 use serde_json::Map;
+use strum::AsStaticRef;
+use strum_macros::AsStaticStr;
 
 /// A JSON RPC Response.
 #[derive(Deserialize, Serialize, Debug)]
@@ -97,18 +104,18 @@ pub struct JsonCommandResponse {
 pub enum JsonRPCError {
     error {
         /// The error code associated with this error.
-        code: JsonRPCErrorCodes,
+        code: i32,
 
         /// A string providing a short description of the error.
         message: String,
 
         /// Additional information about the error.
-        data: String, // FIXME: could be json structured value.
+        data: serde_json::Value,
     },
 }
 
 /// JSON RPC Error codes.
-#[derive(Deserialize, Serialize, Debug)]
+#[derive(Deserialize, Serialize, Debug, AsStaticStr)]
 pub enum JsonRPCErrorCodes {
     /// Parse error.
     ParseError = -32700,
@@ -130,11 +137,12 @@ pub enum JsonRPCErrorCodes {
 
 /// Helper method to format displaydoc errors in JSON RPC 2.0 format.
 pub fn format_error<T: std::fmt::Display + std::fmt::Debug>(e: T) -> String {
-    let data = json!({"server_error": format!("{:?}", e), "details": e.to_string()}).to_string();
+    let data: serde_json::Value =
+        json!({"server_error": format!("{:?}", e), "details": e.to_string()}).into();
     // FIXME: wrap in JsonRPCResponse
     let json_resp = JsonRPCError::error {
-        code: JsonRPCErrorCodes::InternalError,
-        message: "Internal error".to_string(),
+        code: JsonRPCErrorCodes::InternalError as i32,
+        message: JsonRPCErrorCodes::InternalError.as_static().to_string(),
         data,
     };
     json!(json_resp).to_string()
@@ -173,6 +181,15 @@ pub enum JsonCommandResponseV2 {
     get_balance_for_account {
         balance: Balance,
     },
+    send_transaction {
+        transaction_log: TransactionLog,
+    },
+    build_transaction {
+        tx_proposal: TxProposal,
+    },
+    submit_transaction {
+        transaction_log: Option<TransactionLog>,
+    },
     /*
     get_balance_for_subaddress {
         balance: Balance,
@@ -203,15 +220,6 @@ pub enum JsonCommandResponseV2 {
     get_all_addresses_by_account {
         address_ids: Vec<String>,
         address_map: Map<String, serde_json::Value>,
-    },
-    send_transaction {
-        transaction: JsonSubmitResponse,
-    },
-    build_transaction {
-        tx_proposal: StringifiedJsonTxProposal,
-    },
-    submit_transaction {
-        transaction: JsonSubmitResponse,
     },
     get_all_transactions_by_account {
         transaction_log_ids: Vec<String>,
