+++ full-service/src/json_rpc/mod.rs
@@ -9,6 +9,9 @@ pub mod api_v1;
 mod balance;
 pub mod json_rpc_request;
 pub mod json_rpc_response;
+mod transaction_log;
+mod tx_proposal;
+mod unspent_tx_out;
 pub mod wallet;
 mod wallet_status;
